requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-groq
langchain-qdrant
langchain-core
langchain-tavily
langgraph
pypdf
pydantic
python-dotenv
requests
beautifulsoup4 
pandas 
lxml 
html5lib
together
qdrant-client
opik
fastapi
uvicorn
streamlit
assemblyai
audio-recorder-streamlit
cartesia
soundfile
google-generativeai
google-genai 
google-auth 
google-auth-oauthlib 
google-auth-httplib2
tqdm
unstructured
pypdf
pypdf2
unstructured[pdf]
sqlalchemy
alembic
passlib[bcrypt] 

fastapi
uvicorn[standard]
python-multipart
sse-starlette

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Pydantic settings
pydantic-settings

# CORS
fastapi-cors

# Additional dependencies
aiofiles
typing-extensions
-e .